feat(auth): renew the firebase auth token when expired
There are two different ways to detect the token expiration in chat
agent.

1. "auth_revoked" event is sent from the subscribed event stream.
2. Updating the firebase DB fails with status code 401 - unauthorized.

In the first case, the event stream is closed and the entire
initialization logic is rerun. In the second case, the same operation
is retried after re-running the initialization logic.

Just in case both events happen roughly at the same time, I also added
a debounce logic to prevent the initialize() function from running
multiple times in a short time frame.

SO-374 #done

Change-Id: I3e68762cb401701e637ca43c3e5d5c1c769bc725